requirements.txt
streamlit
pdf2image
pillow
opencv-python-headless
pytesseract
pandas
openpyxl